fixed error of floating point conversion

+++ app/Http/Controllers/Controller.php
@@ -57,8 +57,8 @@ class Controller extends BaseController
 
         if ($correct_answers > env('CORRECT_ANSWERS_THRESHOLD'))
         {
-          $difficulty = (float(1.00)-(float($correct_answers)/float($submitted_answers)));
-          Question::where('qid', $qid)->update('difficulty', integer($difficulty));
+          $difficulty = (1.00-((float)$correct_answers/(float)$submitted_answers));
+          Question::where('qid', $qid)->update('difficulty', (integer)$difficulty);
 
           return $base_point*$difficulty;
         }
